BUGFIX: sql-injection, check version
While looking at the code when I was researching a problem I had,
I saw some things that I wanted to change. For example, query strings
where concatenated every time they where used, resulting in a new string
with identical content every time. The way the string where concatenated
made them vulnerable to sql-injection.

First I moved all query-strings to an object, and wrote $TABLE where
the table name will be replaced during the initialization.

+const SQL = Object.freeze({
+  setup:
+    'SELECT version(), quote_ident($1) AS schema, quote_ident($2) AS table',
+  set:
+    'INSERT INTO $TABLE (sess, expire, sid) ' +
+    'SELECT $1, to_timestamp($2), $3 ' +
+    'ON CONFLICT (sid) DO ' + // Needs Postgrsql 9.5 or later for ON CONFLICT DO
+    'UPDATE SET sess=$1, expire=to_timestamp($2) ' +
+    'RETURNING sid',
+  prune:
+    'DELETE FROM $TABLE WHERE expire < to_timestamp($1)',
+  touch:
+    'UPDATE $TABLE SET expire = to_timestamp($1) ' +
+    'WHERE sid = $2 RETURNING sid',
+  destroy:
+    'DELETE FROM $TABLE WHERE sid = $1',
+  get:
+    'SELECT sess FROM $TABLE WHERE sid = $1 AND expire >= to_timestamp($2)'
+});

Then I changed every place they where used:

-    this.query('DELETE FROM ' + this.quotedTable() + ' WHERE expire < to_timestamp($1)', [currentTimestamp()], function (err) {
+    this.query(this.SQL.prune, [currentTimestamp()], function (err) {

-    this.query('SELECT sess FROM ' + this.quotedTable() + ' WHERE sid = $1 AND expire >= to_timestamp($2)', [sid, currentTimestamp()], function (err, data) {
+    this.query(this.SQL.get, [sid, currentTimestamp()], function (err, data) {

-    const query = 'INSERT INTO ' + this.quotedTable() + ' (sess, expire, sid) SELECT $1, to_timestamp($2), $3 ON CONFLICT (sid) DO UPDATE SET sess=$1, expire=to_timestamp($2) RETURNING sid';
-
-    this.query(query, [sess, expireTime, sid], function (err) {
+    this.query(this.SQL.set, [sess, expireTime, sid], function (err) {

-    this.query('DELETE FROM ' + this.quotedTable() + ' WHERE sid = $1', [sid], function (err) {
+    this.query(this.SQL.destroy, [sid], function (err) {

-    this.query(
-      'UPDATE ' + this.quotedTable() + ' SET expire = to_timestamp($1) WHERE sid = $2 RETURNING sid',
-      [expireTime, sid],
-      function (err) { fn(err); }
-    );
+    this.query(this.SQL.touch, [expireTime, sid], function (err) { fn(err); });

In Postgresql key words and unquoted identifiers are case insensitive:
  SeLeCt NuLl
and
  select null
are treated the same. By quoting the identifier it will be treated case
sensitive. Quoting also makes it possible to use a key word as an identifier.

But care must be taken when quoting a string, or it might be vulnerable
for SQL-injection. The function PGStore.prototype.quotedTable was vulnerable
since it didn't handle if the input had double quotes. A schema
name of 'web".session WHERE $1=$1;--' would delete everything in table
web.session every time the prune process runs.

I removed the function:

-  /**
-   * Get the quoted table.
-   *
-   * @return {String} the quoted schema + table for use in queries
-   * @access private
-   */
-
-  PGStore.prototype.quotedTable = function () {
-    let result = '"' + this.tableName + '"';
-
-    if (this.schemaName) {
-      result = '"' + this.schemaName + '".' + result;
-    }
-
-    return result;
-  };

In my refactored version I instead ask the server to quote the identifiers
with the function quote_ident. It will give back a string that only uses
quotes if they are needed. The same query is used to retrive the server
version and both schemaName and tableName:

   'SELECT version(), quote_ident($1) AS schema, quote_ident($2) AS table'

The processed schemaName and tableName are then concatenated to a
fully qualified string:

+      const fqname = (res.schema ? res.schema + '.' : '') + res.table;

And a new object is created and populated with strings with the table
name replacing $TABLE:

+      // Replace $TABLE in query strings with the correct value.
+      this.SQL = Object.keys(SQL).reduce(function (out, key) {
+        out[key] = SQL[key].replace(/\$TABLE/g, fqname);
+        return out;
+      }, {});
+      this.SQL = Object.freeze(this.SQL);

I added a constant with the minium server version.

+  const pgVersionMin = { major: 9, minor: 5 };

and then compare the minium compared version with the current version
of the server, and calls errorLog if the current version is too old.

+      // Extract major and minor version from string
+      let ver = /\s(\d+)\.(\d+)/.exec(res.version);
+      if (!ver) {
+        this.errorLog(Error('Could not get server version'));
+      } else {
+        const min = pgVersionMin;
+        ver = {
+          major: parseInt(ver[1]),
+          minor: parseInt(ver[2])
+        };
+        if (ver.major < min.major || (ver.major === min.major && ver.minor < min.minor)) {
+          this.errorLog(Error(
+            `Too old version of postgresql. Need ${min.major}.${min.minor} or later.`
+          ));
+        }
+      }