BUGFIX: #151 SQL Injection
The function PGStore.prototype.quotedTable was vulnerable since it
didn't correctly handle schema and table names with double quotes.
A specially crafted schema name could have deleted the content of
tables every time the prune process was running.

One way to solve this is to make a call to the server and let the
function quote_ident take care of the quoting. But since that requires a
call to the server at initialization time, and the server might not be
reachable, it quickly becomes complicated.

I solved it by adding two helper functions:

quoteIdent takes care of the double quoting by duplicating double quotes:

  const quoteIdent = function (ident) {
    return '"' + ident.replace(/"/g, '""') + '"';
  };

qualifiedTable returns a quoted string with the schema name if needed.

  const qualifiedTable = function (schema, table) {
    return (schema ? quoteIdent(schema) + '.' : '') + quoteIdent(table);
  };

Instead of concatenatening and quoting the schema and table name at
every query, I changed it to concetenate the table name at
initialization:

  this.qualifiedTableName = qualifiedTable(this.schemaName, this.tableName);

and updated quotedTable to use the calculated value.

  PGStore.prototype.quotedTable = function () {
    return this.qualifiedTableName;
  };

I did not remove quotedTable since it might break some third party code.

I also changed every instance where quotedTable() was called, to use
this.qualifiedTableName directly.

- this.query('DELETE FROM ' + this.quotedTable() + ' WHERE expire < to_timestamp($1)', [currentTimestamp()], function (err) {
+ this.query('DELETE FROM ' + this.qualifiedTableName + ' WHERE expire < to_timestamp($1)', [currentTimestamp()], function (err) {

- this.query('SELECT sess FROM ' + this.quotedTable() + ' WHERE sid = $1 AND expire >= to_timestamp($2)', [sid, currentTimestamp()], function (err, data) {
+ this.query('SELECT sess FROM ' + this.qualifiedTableName + ' WHERE sid = $1 AND expire >= to_timestamp($2)', [sid, currentTimestamp()], function (err, data) {

- const query = 'INSERT INTO ' + this.quotedTable() + ' (sess, expire, sid) SELECT $1, to_timestamp($2), $3 ON CONFLICT (sid) DO UPDATE SET sess=$1, expire=to_timestamp($2) RETURNING sid';
+ const query = 'INSERT INTO ' + this.qualifiedTableName + ' (sess, expire, sid) SELECT $1, to_timestamp($2), $3 ON CONFLICT (sid) DO UPDATE SET sess=$1, expire=to_timestamp($2) RETURNING sid';

- this.query('DELETE FROM ' + this.quotedTable() + ' WHERE sid = $1', [sid], function (err) {
+ this.query('DELETE FROM ' + this.qualifiedTableName + ' WHERE sid = $1', [sid], function (err) {

- 'UPDATE ' + this.quotedTable() + ' SET expire = to_timestamp($1) WHERE sid = $2 RETURNING sid',
+ 'UPDATE ' + this.qualifiedTableName + ' SET expire = to_timestamp($1) WHERE sid = $2 RETURNING sid',